requirements.txt
fastapi
uvicorn
tcgdexsdk
aiohttp
openpyxl
pillow
jinja2